requirements.txt
google-cloud-bigquery==3.17.0
google-auth-oauthlib==1.2.0
google-auth==2.28.0
pandas==2.2.0
python-dotenv==1.0.0
db-dtypes==1.1.0
matplotlib==3.8.0
seaborn==0.13.0
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
streamlit==1.28.1
plotly==5.17.0
